requirements.txt
requests
python-dotenv
pandas
python-binance 